+++ LiteXLoader/src/API/GuiAPI.cpp
@@ -1,13 +1,16 @@
 #include "APIhelp.h"
 #include "GuiAPI.h"
 #include <Kernel/Gui.h>
+#include <iostream>
 using namespace script;
+using namespace std;
 
 //////////////////// Class Definition ////////////////////
 
 ClassDefine<FormClass> FormClassBuilder =
     defineClass<FormClass>("Form")
         .constructor(nullptr)
+        .instanceFunction("setTitle", &FormClass::setTitle)
         .instanceFunction("addLabel", &FormClass::addLabel)
         .instanceFunction("addInput", &FormClass::addInput)
         .instanceFunction("addSwitch", &FormClass::addSwitch)
@@ -20,9 +23,10 @@ ClassDefine<FormClass> FormClassBuilder =
 //////////////////// Classes ////////////////////
 
 FormClass::FormClass()
-    :ScriptClass(ScriptClass::ConstructFromCpp<FormClass>{}),
-     form({{"title",""},{"content",""}})
-{ }
+    :ScriptClass(ScriptClass::ConstructFromCpp<FormClass>{})
+{ 
+    form = JSON_VALUE::parse(R"({ "title":"", "type":"custom_form", "content":[], "buttons":[] })");
+}
 
 
 //生成函数
@@ -53,6 +57,7 @@ Local<Value> FormClass::setTitle(const Arguments& args)
     }
     catch(JSON_VALUE::exception &e){
         ERROR("Fail to Form currect Form string!");
+        ERRPRINT(e.what());
         return Boolean::newBoolean(false);
     }
     CATCH("Fail in setTitle!")
@@ -73,6 +78,7 @@ Local<Value> FormClass::addLabel(const Arguments& args)
     }
     catch(JSON_VALUE::exception &e){
         ERROR("Fail to Form currect Form string!");
+        ERRPRINT(e.what());
         return Boolean::newBoolean(false);
     }
     CATCH("Fail in addLabel!")
@@ -99,6 +105,7 @@ Local<Value> FormClass::addInput(const Arguments& args)
     }
     catch(JSON_VALUE::exception &e){
         ERROR("Fail to Form currect Form string!");
+        ERRPRINT(e.what());
         return Boolean::newBoolean(false);
     }
     CATCH("Fail in addInput!")
@@ -113,7 +120,7 @@ Local<Value> FormClass::addSwitch(const Arguments& args)
 
     try{
         JSON_VALUE itemAdd;
-        itemAdd["type"] = "input";
+        itemAdd["type"] = "toggle";
         itemAdd["text"] = args[0].toStr();
         if(args.size() >= 2)
             itemAdd["default"] = args[1].asBoolean().value();
@@ -123,6 +130,7 @@ Local<Value> FormClass::addSwitch(const Arguments& args)
     }
     catch(JSON_VALUE::exception &e){
         ERROR("Fail to Form currect Form string!");
+        ERRPRINT(e.what());
         return Boolean::newBoolean(false);
     }
     CATCH("Fail in addSwitch!")
@@ -148,6 +156,7 @@ Local<Value> FormClass::addDropdown(const Arguments& args)
     }
     catch(JSON_VALUE::exception &e){
         ERROR("Fail to Form currect Form string!");
+        ERRPRINT(e.what());
         return Boolean::newBoolean(false);
     }
     CATCH("Fail in addDropdown!")
@@ -191,6 +200,7 @@ Local<Value> FormClass::addSlider(const Arguments& args)
     }
     catch(JSON_VALUE::exception &e){
         ERROR("Fail to Form currect Form string!");
+        ERRPRINT(e.what());
         return Boolean::newBoolean(false);
     }
     CATCH("Fail in addSlider!")
@@ -223,6 +233,7 @@ Local<Value> FormClass::addStepSlider(const Arguments& args)
     }
     catch(JSON_VALUE::exception &e){
         ERROR("Fail to Form currect Form string!");
+        ERRPRINT(e.what());
         return Boolean::newBoolean(false);
     }
     CATCH("Fail in addStepSlider!")