+++ LiteXLoader/src/API/LxlAPI.h
@@ -3,7 +3,4 @@
 using namespace script;
 
 Local<Value> LxlGetVersion(const Arguments& args);
-//Local<Value> LxlLoadPlugin(const Arguments& args);
 Local<Value> LxlListPlugins(const Arguments& args);
-
-Local<Value> LxlImport(const Arguments& args);