+++ LiteXLoader/src/Kernel/Packet.h
@@ -9,10 +9,10 @@ void* Raw_CreatePacket(int type);
 bool Raw_SendPacket(Player *player,void *packet);
 
 int Raw_SendFormPacket(Player* player, const std::string &data);
-bool Raw_SendTransferPacket(Player* player, const string& address, short port);
-bool Raw_SendSetDisplayObjectivePacket(Player* player, const string& title, const string& name);
-bool Raw_SendSetScorePacket(Player* player, char type, const vector<ScorePacketInfo>& data);
+bool Raw_SendTransferPacket(Player* player, const std::string& address, short port);
+bool Raw_SendSetDisplayObjectivePacket(Player* player, const std::string& title, const std::string& name);
+bool Raw_SendSetScorePacket(Player* player, char type, const std::vector<ScorePacketInfo>& data);
 
-bool Raw_SendBossEventPacket(Player* player, string name, float percent, int type);
+bool Raw_SendBossEventPacket(Player* player, std::string name, float percent, int type);
 
 Player* Raw_GetPlayerFromPacket(ServerNetworkHandler* handler, NetworkIdentifier* id, void* packet);