+++ LiteXLoader/src/Kernel/i18n.cpp
@@ -1,0 +1,11 @@
+#include "i18n.h"
+#include "Global.h"
+#include <string>
+using namespace std;
+
+LangPack LangP;
+
+void InitI18n(const string &langPack)
+{
+	LangP.load(langPack);
+}